Improve canonicalize_email() to make it more correct and faster.
The function canonicalize_email() had a bug making it returns the wrong
email for the first call. This reason is the variable email used both
for the loop and for the argument (the loop overriding the argument).

Since canon_map is always used, the branch to populate it is lazily
is not adding anything. The code to populate the dictionnary is taken
out of the function, fixing the bug and the performance issue.